backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
openreview-py>=1.58.0
httpx>=0.27.2
arxiv>=2.1.3
numpy>=2.0.0
scikit-learn>=1.5.2
umap-learn>=0.5.6
python-dotenv>=1.0.1
pydantic>=2.9.2
cachetools>=5.5.0
aiofiles>=24.1.0
PyYAML>=6.0
